backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.5
httpx>=0.24.1
pytest>=7.4.0 
openai>=1.0.0
pyyaml>=6.0